pyproject.toml
[tool.poetry]
name = "py-aiger"
readme="README.md"
version = "7.0.0"
repository = "https://github.com/mvcisback/py-aiger"
description = "A python library for manipulating sequential and-inverter gates."
authors = ["Marcell Vazquez-Chanlatte <marcell.vc@eecs.berkeley.edu>"]
license = "MIT"
packages = [
    { include = "aiger" },
]

[tool.poetry.dependencies]
python = "^3.7"
attrs = ">=23"
bidict = "^0.22.0"
funcy = "^1.12"
pyrsistent = "^0.19.0"
sortedcontainers = "^2.3.0"

[tool.poetry.dev-dependencies]
hypothesis_cfg = {git = "https://github.com/mvcisback/hypothesis-cfg.git"}
hypothesis = "^6"
pytest-xdist = "^3"
py-aiger-ptltl = "^3.1.0"
parsimonious = "^0.10"

[tool.poetry.group.dev.dependencies]
pytest-timeout = "^2.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
